+++ src/main/java/online/kingdomkeys/kingdomkeys/world/dimension/ModDimensions.java
@@ -22,6 +22,7 @@ public class ModDimensions {
     public static final ResourceKey<Level> STATION_OF_SORROW = ResourceKey.create(Registries.DIMENSION, KingdomKeys.rl(Strings.stationOfSorrow));
     public static final ResourceKey<Level> CASTLE_OBLIVION = ResourceKey.create(Registries.DIMENSION, KingdomKeys.rl(Strings.castleOblivion));
     public static final ResourceKey<Level> DAYBREAK_TOWN = ResourceKey.create(Registries.DIMENSION, KingdomKeys.rl(Strings.daybreakTown));
+    public static final ResourceKey<Level> DESTINY_ISLANDS = ResourceKey.create(Registries.DIMENSION, KingdomKeys.rl(Strings.destinyIslands));
     public static final ResourceKey<Level> WORLDMAP = ResourceKey.create(Registries.DIMENSION, KingdomKeys.rl(Strings.worldmap));
 
     public static final DeferredRegister<MapCodec<? extends ChunkGenerator>> CHUNK_GENERATORS = DeferredRegister.create(Registries.CHUNK_GENERATOR,  KingdomKeys.MODID);